# internal/swagger/swagger.json
{
    "schemes": [
        "http",
        "https"
    ],
    "swagger": "2.0",
    "info": {
        "description": "JD self-operated product price monitoring service based on Gin, GORM and Wire DI.",
        "title": "ginadmin",
        "contact": {},
        "version": "v10.1.0"
    },
    "basePath": "/",
    "paths": {
        "/api/v1/shop/alerts": {
            "get": {
                "tags": [
                    "ShopAPI"
                ],
                "summary": "Query price alerts",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "pagination index",
                        "name": "current",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "pagination size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "product ID",
                        "name": "productID",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pending, sent, or failed",
                        "name": "sendStatus",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "start time",
                        "name": "startTime",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "end time",
                        "name": "endTime",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/schema.PriceAlert"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
        "/api/v1/shop/categories": {
            "get": {
                "tags": [
                    "ShopAPI"
                ],
                "summary": "Query JD categories",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "pagination index",
                        "name": "current",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "pagination size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "category name",
                        "name": "name",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "enabled or disabled",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/schema.JDCategory"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "ShopAPI"
                ],
                "summary": "Create JD category",
                "parameters": [
                    {
                        "description": "category",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.JDCategoryForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.JDCategory"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
        "/api/v1/shop/categories/{id}": {
            "get": {
                "tags": [
                    "ShopAPI"
                ],
                "summary": "Get JD category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.JDCategory"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "ShopAPI"
                ],
                "summary": "Update JD category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "category",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/schema.JDCategoryForm"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "409": {
                        "description": "Conflict",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            },
            "delete": {
                "tags": [
                    "ShopAPI"
                ],
                "summary": "Delete JD category",
                "parameters": [
                    {
                        "type": "string",
                        "description": "category ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "404": {
                        "description": "Not Found",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
        "/api/v1/shop/jobs": {
            "get": {
                "tags": [
                    "ShopAPI"
                ],
                "summary": "Query shop job runs",
                "parameters": [
                    {
                        "type": "integer",
                        "default": 1,
                        "description": "pagination index",
                        "name": "current",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "integer",
                        "default": 10,
                        "description": "pagination size",
                        "name": "pageSize",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "job type",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "job status",
                        "name": "status",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/schema.JobRun"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
        "/api/v1/shop/jobs/{type}/run": {
            "post": {
                "tags": [
                    "ShopAPI"
                ],
                "summary": "Trigger an asynchronous shop job",
                "parameters": [
                    {
                        "type": "string",
                        "description": "discover, public-scan, or checkout-sample",
                        "name": "type",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.JobRun"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "400": {
                        "description": "Bad Request",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
        "/api/v1/shop/notifications/test": {
            "post": {
                "tags": [
                    "ShopAPI"
                ],
                "summary": "Send a ServerChan test notification",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.NotificationTestResult"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
...
        "/api/v1/shop/session": {
            "get": {
                "tags": [
                    "ShopAPI"
                ],
                "summary": "Get JD browser session status",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/util.ResponseResult"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/schema.SessionStatus"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "500": {
                        "description": "Internal Server Error",
                        "schema": {
                            "$ref": "#/definitions/util.ResponseResult"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "schema.JDCategory": {
            "type": "object",
            "properties": {
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "last_discovered_at": {
                    "type": "string"
                },
                "last_discovery_error": {
                    "type": "string"
                },
                "last_discovery_status": {
                    "type": "string"
                },
                "max_pages": {
                    "type": "integer"
                },
                "name": {
                    "type": "string"
                },
                "source_url": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
        "schema.JDCategoryForm": {
            "type": "object",
            "required": [
                "max_pages",
                "name",
                "source_url",
                "status"
            ],
            "properties": {
                "max_pages": {
                    "type": "integer",
                    "maximum": 50,
                    "minimum": 1
                },
                "name": {
                    "type": "string",
                    "maxLength": 128
                },
                "source_url": {
                    "type": "string",
                    "maxLength": 768
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "enabled",
                        "disabled"
                    ]
                }
            }
        },
...
        "schema.NotificationTestResult": {
            "type": "object",
            "properties": {
                "provider": {
                    "type": "string"
                },
                "response": {
                    "type": "string"
                }
            }
        },
        "schema.PriceAlert": {
            "type": "object",
            "properties": {
                "baseline_price_fen": {
                    "type": "integer"
                },
                "baseline_price_yuan": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "current_price_fen": {
                    "type": "integer"
                },
                "current_price_yuan": {
                    "type": "string"
                },
                "drop_basis_points": {
                    "type": "integer"
                },
                "id": {
                    "type": "string"
                },
                "last_error": {
                    "type": "string"
                },
                "product": {
                    "$ref": "#/definitions/schema.JDProduct"
                },
                "product_id": {
                    "type": "string"
                },
                "provider_response": {
                    "type": "string"
                },
                "send_attempts": {
                    "type": "integer"
                },
                "send_status": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "string"
                },
                "trigger_sample_id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "util.ResponseResult": {
            "type": "object",
            "properties": {
                "data": {},
                "error": {
                    "$ref": "#/definitions/errors.Error"
                },
                "success": {
                    "type": "boolean"
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}